+++ user.go
@@ -4,6 +4,7 @@ import (
 	"bytes"
 	"encoding/json"
 	"errors"
+
 	apiLib "github.com/hornbill/goApiLib"
 )
 
@@ -38,6 +39,9 @@ func userCreate(hIF *apiLib.XmlmcInstStruct, user *userWorkingDataStruct, buffer
 	//-- Password is base64 encoded already in process_data
 	hIF.SetParam("password", user.Account.Password)
 	hIF.SetParam("userType", user.Account.UserType)
+	if user.Account.Enable2FA != "" {
+		hIF.SetParam("enable2fa", user.Account.Enable2FA)
+	}
 	if user.Account.FirstName != "" {
 		hIF.SetParam("firstName", user.Account.FirstName)
 	}
@@ -84,8 +88,21 @@ func userCreate(hIF *apiLib.XmlmcInstStruct, user *userWorkingDataStruct, buffer
 	if user.Account.CountryCode != "" {
 		hIF.SetParam("countryCode", user.Account.CountryCode)
 	}
-	//hIF.SetParam("notifyEmail", "")
-	//hIF.SetParam("notifyTextMessage", "")
+	if user.Account.DisableDirectLogin != "" ||
+		user.Account.DisableDirectLoginPasswordReset != "" ||
+		user.Account.DisableDevicePairing != "" {
+		hIF.OpenElement("securityOptions")
+		if user.Account.DisableDirectLogin != "" {
+			hIF.SetParam("disableDirectLogin", user.Account.DisableDirectLogin)
+		}
+		if user.Account.DisableDirectLoginPasswordReset != "" {
+			hIF.SetParam("disableDirectLoginPasswordReset", user.Account.DisableDirectLoginPasswordReset)
+		}
+		if user.Account.DisableDevicePairing != "" {
+			hIF.SetParam("disableDevicePairing", user.Account.DisableDevicePairing)
+		}
+		hIF.CloseElement("securityOptions")
+	}
 
 	//-- Dry Run
 	if Flags.configDryRun {
@@ -130,6 +147,9 @@ func userUpdate(hIF *apiLib.XmlmcInstStruct, user *userWorkingDataStruct, buffer
 
 	if user.Account.UserType != "" {
 		hIF.SetParam("userType", user.Account.UserType)
+	}
+	if user.Account.Enable2FA != "" {
+		hIF.SetParam("enable2fa", user.Account.Enable2FA)
 	}
 	if user.Account.Name != "" {
 		hIF.SetParam("name", user.Account.Name)
@@ -181,8 +201,13 @@ func userUpdate(hIF *apiLib.XmlmcInstStruct, user *userWorkingDataStruct, buffer
 	if user.Account.CountryCode != "" {
 		hIF.SetParam("countryCode", user.Account.CountryCode)
 	}
-	//hIF.SetParam("notifyEmail", "")
-	//hIF.SetParam("notifyTextMessage", "")
+	if user.Account.UpdateSecOptions {
+		hIF.OpenElement("securityOptions")
+		hIF.SetParam("disableDirectLogin", user.Account.DisableDirectLogin)
+		hIF.SetParam("disableDirectLoginPasswordReset", user.Account.DisableDirectLoginPasswordReset)
+		hIF.SetParam("disableDevicePairing", user.Account.DisableDevicePairing)
+		hIF.CloseElement("securityOptions")
+	}
 	var XMLSTRING = hIF.GetParam()
 	//-- Dry Run
 	if Flags.configDryRun {